pyproject.toml
[project]
name = "bugred-cicd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "allure-pytest>=2.14.3",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "webdriver-manager>=4.0.2",
]
